refactor: [FE] PlayBarUtil 파일명 변경
- WidthUtil로 이름 변경했습니다.
- 너비를 구하는 함수들이 모여있습니다.
