+++ frontend/src/common/util/index.js
@@ -1,7 +1,7 @@
 export * as EventUtil from './EventUtil';
 export * as FileUtil from './FileUtil';
 export * as AudioUtil from './AudioUtil';
-export * as PlayBarUtil from './PlayBarUtil';
+export * as WidthUtil from './WidthUtil';
 export * as MarkerEventUtil from './MarkerEventUtil';
 export * as CopyUtil from './CopyUtil';
 export * as SectionEffectListUtil from './SectionEffectListUtil';
